+++ src/modules/orders/orderController.js
@@ -1,33 +1,33 @@
 
-const orderItems = require('../model/orderItems')
+const OrderItems = require('../../model/orderItems')
 
-exports.GetOrderItems = async (req, res) => {
+exports.getOrderItems = async (req, res) => {
     try {
-        const orderData = await orderItems.query();
+        const orderData = await OrderItems.query();
         res.status(200).json({ message: "Successfully Fetched Order Items", orderData })
     } catch (error) {
         res.status(500).json({ message: "Error in Fetching Order Items", error: error.message })
     }
 }
 
-exports.CreateOrderItems = async (req, res) => {
+exports.createOrderItem = async (req, res) => {
     const orderData = req.body;
     if (!orderData.orderId || !orderData.menuItemId || !orderData.quantity || !orderData.itemPrice || !orderData.total) {
         res.status(400).json({ message: 'Fill all the fields' })
     }
     try {
-        const datas = await orderItems.query().insert(orderData)
+        const datas = await OrderItems.query().insert(orderData)
         res.status(200).json({ message: 'Successfully Created Order Items', datas })
     } catch (error) {
         res.status(500).json({ message: "Error in Creating Order Items", error: error.message })
     }
 }
 
-exports.UpdateOrderItems = async (req, res) => {
+exports.updateOrderItem = async (req, res) => {
     const { id } = req.params;
     const updatedData = req.body;
     try {
-        const updated = await orderItems.query().patchAndFetchById(id, updatedData);
+        const updated = await OrderItems.query().patchAndFetchById(id, updatedData);
         if (!updated) {
             return res.status(404).json({ message: 'Order Item not found' });
         }
@@ -37,10 +37,10 @@ exports.UpdateOrderItems = async (req, res) => {
     }
 };
 
-exports.DeleteOrderItems = async (req, res) => {
+exports.deleteOrderItem = async (req, res) => {
     const { id } = req.params;
     try {
-        const deleted = await orderItems.query().deleteById(id);
+        const deleted = await OrderItems.query().deleteById(id);
 
         if (deleted === 0) {
             return res.status(404).json({ message: 'Order Item not found' });
